requirements.txt
langchain>=0.2.12
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
faiss-cpu>=1.8.0.post1
PyMuPDF>=1.24.7
langchain-google-genai>=2.0.1
langgraph>=0.2.26
python-dotenv>=1.0.1